moved sqsErrors and decodeMessage into commons package
