+++ packages/commons/src/sqs/index.ts
@@ -1,1 +1,3 @@
 export * as SQS from "./sqs.js";
+export * from "./utils.js";
+export * from "./sqsErrors.js";
