+++ packages/commons/src/sqs/sqsErrors.ts
