+++ packages/commons/src/sqs/utils.ts
@@ -1,6 +1,6 @@
 import { SQS } from "pagopa-interop-kpi-commons";
 import { S3BodySchema } from "pagopa-interop-kpi-models";
-import { decodeSQSEventMessageError } from "./errors.js";
+import { decodeSQSEventMessageError } from "./sqsErrors.js";
 
 export function decodeSQSEventMessage(message: SQS.Message): string {
   try {
